applied patch #1680556 - fixes to allow zip2exe & Makensisw to build in winelib
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS@5008 212acab6-be3b-0410-9dea-997c60f758d6

+++ trunk/Contrib/Makensisw/jnetlib/asyncdns.cpp
@@ -30,7 +30,7 @@ JNL_AsyncDNS::~JNL_AsyncDNS()
   }
 }
 
-unsigned long WINAPI JNL_AsyncDNS::_threadfunc(LPVOID _d)
+DWORD WINAPI JNL_AsyncDNS::_threadfunc(LPVOID _d)
 {
   JNL_AsyncDNS *_this=(JNL_AsyncDNS*)_d;
   int nowinsock=JNL::open_socketlib();
